🗑️🚚 Deprecate fc.dedup in favor of fc.clone
Related to #530
