+++ test/e2e/StateFullArbitraries.spec.ts
@@ -25,9 +25,9 @@ describe(`StateFullArbitraries (seed: ${seed})`, () => {
       fc.assert(fc.property(cloneableWithCount(data), cloneableWithCount(data), () => {}));
       expect(data.counter).toEqual(0);
     });
-    it('fc.dedup', () => {
+    it('fc.clone', () => {
       const data = { counter: 0 };
-      fc.assert(fc.property(fc.dedup(cloneableWithCount(data), 3), () => {}));
+      fc.assert(fc.property(fc.clone(cloneableWithCount(data), 3), () => {}));
       expect(data.counter).toEqual(0);
     });
     it('fc.tuple', () => {
@@ -100,10 +100,10 @@ describe(`StateFullArbitraries (seed: ${seed})`, () => {
       expect(nonClonedDetected).toBe(false);
       expect(status.counterexample![1]!.size()).toEqual(1);
     });
-    it('fc.dedup', () => {
+    it('fc.clone', () => {
       let nonClonedDetected = false;
       const status = fc.check(
-        fc.property(fc.integer(), fc.dedup(fc.context(), 3), fc.integer(), (a, ctxs, b) => {
+        fc.property(fc.integer(), fc.clone(fc.context(), 3), fc.integer(), (a, ctxs, b) => {
           for (const ctx of ctxs) {
             nonClonedDetected = nonClonedDetected || ctx.size() !== 0;
             ctx.log('logging stuff');
