+++ test/type/index.test-d.ts
@@ -125,13 +125,13 @@ expectType<{ a: fc.Arbitrary<number>; b: fc.Arbitrary<unknown> }>(
   }))
 ); // TODO Typings should be improved: referencing an undefined key should failed
 
-// dedup arbitrary
-expectType<fc.Arbitrary<[]>>(fc.dedup(fc.nat(), 0));
-expectType<fc.Arbitrary<[number]>>(fc.dedup(fc.nat(), 1));
-expectType<fc.Arbitrary<[number, number]>>(fc.dedup(fc.nat(), 2));
-expectType<fc.Arbitrary<[number, number, number]>>(fc.dedup(fc.nat(), 3));
-expectType<fc.Arbitrary<[number, number, number, number]>>(fc.dedup(fc.nat(), 4));
-expectType<fc.Arbitrary<number[]>>(fc.dedup(fc.nat(), 5)); // TODO Typings should be improved: handle any number of values
+// clone arbitrary
+expectType<fc.Arbitrary<[]>>(fc.clone(fc.nat(), 0));
+expectType<fc.Arbitrary<[number]>>(fc.clone(fc.nat(), 1));
+expectType<fc.Arbitrary<[number, number]>>(fc.clone(fc.nat(), 2));
+expectType<fc.Arbitrary<[number, number, number]>>(fc.clone(fc.nat(), 3));
+expectType<fc.Arbitrary<[number, number, number, number]>>(fc.clone(fc.nat(), 4));
+expectType<fc.Arbitrary<number[]>>(fc.clone(fc.nat(), 5)); // TODO Typings should be improved: handle any number of values
 
 // func arbitrary
 expectType<fc.Arbitrary<() => number>>(fc.func(fc.nat()));
